+++ RELEASE_NOTES.md
@@ -6,3 +6,4 @@ for usage.
 Implemented:
  - [Widgets API](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget/createwidget)
  - [Widget Templates](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget-template/createwidgettemplate)
+ - Get Theme Regions